Added Templating to Glossary
I want to start adding Jupyter Book terms since we will be using it a lot for the rest of the class. Would it make sense to create a new page that is strictly for Jupyter Book help/terms and keep this one as strictly git/github/bash or keep all of the terms onto one page?